+++ marshal/host.go
@@ -30,19 +30,19 @@ func Host(cmdOutput []byte) model.Host {
 	m.Hostname = strings.TrimSpace(data["Hostname"])
 	m.CPUModel = strings.TrimSpace(data["CPUModel"])
 	m.CPUFrequency = strings.TrimSpace(data["CPUFrequency"])
-	m.CPUSockets = trimParseInt(data["CPUSockets"])
-	m.CPUCores = trimParseInt(data["CPUCores"])
-	m.CPUThreads = trimParseInt(data["CPUThreads"])
-	m.ThreadsPerCore = trimParseInt(data["ThreadsPerCore"])
-	m.CoresPerSocket = trimParseInt(data["CoresPerSocket"])
+	m.CPUSockets = TrimParseInt(data["CPUSockets"])
+	m.CPUCores = TrimParseInt(data["CPUCores"])
+	m.CPUThreads = TrimParseInt(data["CPUThreads"])
+	m.ThreadsPerCore = TrimParseInt(data["ThreadsPerCore"])
+	m.CoresPerSocket = TrimParseInt(data["CoresPerSocket"])
 	m.HardwareAbstraction = strings.TrimSpace(data["HardwareAbstraction"])
 	m.HardwareAbstractionTechnology = strings.TrimSpace(data["HardwareAbstractionTechnology"])
 	m.Kernel = strings.TrimSpace(data["Kernel"])
 	m.KernelVersion = strings.TrimSpace(data["KernelVersion"])
 	m.OS = strings.TrimSpace(data["OS"])
 	m.OSVersion = strings.TrimSpace(data["OSVersion"])
-	m.MemoryTotal = trimParseFloat64(data["MemoryTotal"])
-	m.SwapTotal = trimParseFloat64(data["SwapTotal"])
+	m.MemoryTotal = TrimParseFloat64(data["MemoryTotal"])
+	m.SwapTotal = TrimParseFloat64(data["SwapTotal"])
 
 	return m
 }
