+++ marshal/oracle/tablespaces.go
@@ -13,37 +13,36 @@
 // You should have received a copy of the GNU General Public License
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-package marshal
+package oracle
 
 import (
 	"bufio"
-	"strconv"
 	"strings"
 
+	"github.com/ercole-io/ercole-agent/marshal"
 	"github.com/ercole-io/ercole/model"
 )
 
-// Backups marshals a backup output list into a struct.
-func Backups(cmdOutput []byte) []model.OracleDatabaseBackup {
-	backups := []model.OracleDatabaseBackup{}
-
+// Tablespaces returns information about database tablespaces extracted
+// from the tablespaces fetcher command output.
+func Tablespaces(cmdOutput []byte) []model.OracleDatabaseTablespace {
+	tablespaces := []model.OracleDatabaseTablespace{}
 	scanner := bufio.NewScanner(strings.NewReader(string(cmdOutput)))
 
 	for scanner.Scan() {
-		backup := new(model.OracleDatabaseBackup)
+		tablespace := new(model.OracleDatabaseTablespace)
 		line := scanner.Text()
 		splitted := strings.Split(line, "|||")
-		if len(splitted) == 5 {
-			backup.BackupType = strings.TrimSpace(splitted[0])
-			backup.Hour = strings.TrimSpace(splitted[1])
+		if len(splitted) == 9 {
+			tablespace.Name = strings.TrimSpace(splitted[3])
+			tablespace.MaxSize = marshal.TrimParseFloat64(splitted[4])
+			tablespace.Total = marshal.TrimParseFloat64(splitted[5])
+			tablespace.Used = marshal.TrimParseFloat64(splitted[6])
+			tablespace.UsedPerc = marshal.TrimParseFloat64(splitted[7])
+			tablespace.Status = strings.TrimSpace(splitted[8])
 
-			weekDays := strings.TrimSpace(splitted[2])
-			backup.WeekDays = strings.Split(weekDays, ",")
-
-			backup.AvgBckSize, _ = strconv.ParseFloat(splitted[3], 64)
-			backup.Retention = strings.TrimSpace(splitted[4])
-			backups = append(backups, *backup)
+			tablespaces = append(tablespaces, *tablespace)
 		}
 	}
-	return backups
+	return tablespaces
 }
